@@ -1,5 +1,11 @@
 import React from 'react'
+import { useNavigateMfe } from '../hooks/useNavigateMfe';
 
-const Home = () => (<h1>Welcome to My Microfrontend!</h1>)
+const Home = () => {
+const {insideAbsolute} = useNavigateMfe()
+return <div>
+<h1>Welcome to My Microfrontend!</h1>
+<button onClick={()=> {insideAbsolute('/microfrontend/subroute')}}>Ir a subroute</button>
+</div>}
 
 export default Home;
